Update minSdk version to the actual minimal supported version
Since methods were being directly called that needed sdk 21,
the actual minimal version that's supported is sdk 21.

By bumping the sdk to 21 multiDex can also be enabled without
the support library.

Further more android studio was giving warnings about there being newer,
versions of the support and design libraries, so they are also updated.

Lastly in the latest version of android studio the google repository,
needed to be added to the root build.gradle file for the project to sync
correctly.

fixes #47, #48