Now the proveCommitedNumberIsSquare does not function Something goes wrong when translating between the implementations of the two papers.
The problem is that on RangeProofProver.kt:74, we SHOULD pass w*w to make cDPrime equal to the commitment created by those parameters, this can be seen better in proveCommitedNumberIsSquare:165.
E = g^w h^r1 modN
But we're passing values such that
E = g^(w*w) h^r1 modN

So that's doomed to fail.
But the thing is, the proveCommittedNumberIsSquare should create the intermediate step, and thus DOES need access to w. It should create an F = g*2 h^r_2
such that E == F^x h^r3.

But I don't know how to fix that, so it's still broken.

The value of r3 is created as r1 - r2*w
I'm not sure why, but because we use it when doing F^x h^r3, it's probably relevant.

Then we're passing all these values  to the proveTwoCommittedIntegersAreEqual, which fails because of an innate property: In that function, E != y1 (one is calculated in the function as g1^x h1^r1, the other one is given as an argument)
But that's correct because it's actually g1^(x^2), because of the square thingy.

I don't even know anymore.