Attempt to remove the hardcoded zkp
Ui for this is setup, and we generate a proof.
However we need to save the particular private result.
We need to be able to retrieve this on validation.
However the block creation happens on a grpc thread.
Therefore accessing realm from there is not possible right now.

But we rely on the seq_no of the block to retrieve the private result
when verification happens.

Project currently doesn't compile